Cleaned up the storage package
