+++ storage/stow_store.go
@@ -7,6 +7,15 @@ import (
 	"sync"
 	"time"
 
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	s32 "github.com/aws/aws-sdk-go/service/s3"
+	"github.com/graymeta/stow/azure"
+	"github.com/graymeta/stow/google"
+	"github.com/graymeta/stow/local"
+	"github.com/graymeta/stow/oracle"
+	"github.com/graymeta/stow/s3"
+	"github.com/graymeta/stow/swift"
+
 	"github.com/lyft/flytestdlib/contextutils"
 	"github.com/lyft/flytestdlib/logger"
 	"github.com/lyft/flytestdlib/promutils/labeled"
@@ -23,6 +32,54 @@ const (
 	FailureTypeLabel contextutils.Key = "failure_type"
 )
 
+var fQNFn = map[string]func(string) DataReference{
+	s3.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("s3://%s", bucket))
+	},
+	google.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("gs://%s", bucket))
+	},
+	oracle.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("os://%s", bucket))
+	},
+	swift.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("sw://%s", bucket))
+	},
+	azure.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("afs://%s", bucket))
+	},
+	local.Kind: func(bucket string) DataReference {
+		return DataReference(fmt.Sprintf("file://%s", bucket))
+	},
+}
+
+// Checks if the error is AWS S3 bucket not found error
+func awsBucketIsNotFound(err error) bool {
+	if IsNotFound(err) {
+		return true
+	}
+
+	if awsErr, errOk := errs.Cause(err).(awserr.Error); errOk {
+		return awsErr.Code() == s32.ErrCodeNoSuchBucket
+	}
+
+	return false
+}
+
+// Checks if the error is AWS S3 bucket already exists error.
+func awsBucketAlreadyExists(err error) bool {
+	if IsExists(err) {
+		return true
+	}
+
+	if awsErr, errOk := errs.Cause(err).(awserr.Error); errOk {
+		return awsErr.Code() == s32.ErrCodeBucketAlreadyOwnedByYou
+	}
+
+	return false
+}
+
+// Metrics for Stow store
 type stowMetrics struct {
 	BadReference labeled.Counter
 	BadContainer labeled.Counter
@@ -37,6 +94,7 @@ type stowMetrics struct {
 	WriteLatency labeled.StopWatch
 }
 
+// Metadata that will be returned
 type StowMetadata struct {
 	exists bool
 	size   int64
@@ -237,4 +295,62 @@ func NewStowRawStore(baseContainerFQN DataReference, loc stow.Location, enableDy
 	}
 	self.baseContainer = container
 	return self, nil
+}
+
+// Constructor for the StowRawStore
+func newStowRawStore(cfg *Config, metricsScope promutils.Scope) (RawStore, error) {
+	if cfg.InitContainer == "" {
+		return nil, fmt.Errorf("initContainer is required even with `enable-multicontainer`")
+	}
+
+	var cfgMap stow.ConfigMap
+	var kind string
+	if cfg.Stow != nil {
+		kind = cfg.Stow.Kind
+		cfgMap = cfg.Stow.Config
+	} else {
+		logger.Warnf(context.TODO(), "stow configuration section missing, defaulting to legacy s3/minio connection config")
+		// This is for supporting legacy configurations which configure S3 via connection config
+		kind = s3.Kind
+		cfgMap = legacyS3ConfigMap(cfg.Connection)
+	}
+
+	fn, ok := fQNFn[kind]
+	if !ok {
+		return nil, errs.Errorf("unsupported stow.kind [%s], add support in flytestdlib?", kind)
+	}
+
+	loc, err := stow.Dial(kind, cfgMap)
+	if err != nil {
+		return emptyStore, fmt.Errorf("unable to configure the storage for %s. Error: %v", kind, err)
+	}
+
+	return NewStowRawStore(fn(cfg.InitContainer), loc, cfg.MultiContainerEnabled, metricsScope)
+}
+
+func legacyS3ConfigMap(cfg ConnectionConfig) stow.ConfigMap {
+	// Non-nullable fields
+	stowConfig := stow.ConfigMap{
+		s3.ConfigAuthType: cfg.AuthType,
+		s3.ConfigRegion:   cfg.Region,
+	}
+
+	// Fields that differ between minio and real S3
+	if endpoint := cfg.Endpoint.String(); endpoint != "" {
+		stowConfig[s3.ConfigEndpoint] = endpoint
+	}
+
+	if accessKey := cfg.AccessKey; accessKey != "" {
+		stowConfig[s3.ConfigAccessKeyID] = accessKey
+	}
+
+	if secretKey := cfg.SecretKey; secretKey != "" {
+		stowConfig[s3.ConfigSecretKey] = secretKey
+	}
+
+	if disableSsl := cfg.DisableSSL; disableSsl {
+		stowConfig[s3.ConfigDisableSSL] = "True"
+	}
+
+	return stowConfig
 }
