+++ storage/stow_store_test.go
@@ -7,15 +7,21 @@ import (
 	"io"
 	"io/ioutil"
 	"net/url"
+	"os"
+	"path/filepath"
 	"testing"
 	"time"
 
+	"github.com/graymeta/stow/google"
+	"github.com/graymeta/stow/local"
 	"github.com/pkg/errors"
 
 	"github.com/graymeta/stow"
 	"github.com/stretchr/testify/assert"
 
+	"github.com/lyft/flytestdlib/config"
 	"github.com/lyft/flytestdlib/contextutils"
+	"github.com/lyft/flytestdlib/internal/utils"
 	"github.com/lyft/flytestdlib/promutils"
 	"github.com/lyft/flytestdlib/promutils/labeled"
 )
@@ -212,4 +218,158 @@ func TestStowStore_ReadRaw(t *testing.T) {
 		_, err = s.ReadRaw(context.TODO(), "s3://bad-container/path")
 		assert.Error(t, err)
 	})
+}
+
+func TestNewLocalStore(t *testing.T) {
+	labeled.SetMetricKeys(contextutils.ProjectKey, contextutils.DomainKey, contextutils.WorkflowIDKey, contextutils.TaskIDKey)
+	t.Run("Valid config", func(t *testing.T) {
+		testScope := promutils.NewTestScope()
+		store, err := newStowRawStore(&Config{
+			Stow: &StowConfig{
+				Kind: local.Kind,
+				Config: map[string]string{
+					local.ConfigKeyPath: "./",
+				},
+			},
+			InitContainer: "testdata",
+		}, testScope.NewSubScope("x"))
+
+		assert.NoError(t, err)
+		assert.NotNil(t, store)
+
+		// Stow local store expects the full path after the container portion (looks like a bug to me)
+		rc, err := store.ReadRaw(context.TODO(), DataReference("file://testdata/config.yaml"))
+		assert.NoError(t, err)
+		if assert.NotNil(t, rc) {
+			assert.NoError(t, rc.Close())
+		}
+	})
+
+	t.Run("Invalid config", func(t *testing.T) {
+		testScope := promutils.NewTestScope()
+		_, err := newStowRawStore(&Config{}, testScope)
+		assert.Error(t, err)
+	})
+
+	t.Run("Initialize container", func(t *testing.T) {
+		testScope := promutils.NewTestScope()
+		tmpDir, err := ioutil.TempDir("", "stdlib_local")
+		assert.NoError(t, err)
+
+		stats, err := os.Stat(tmpDir)
+		assert.NoError(t, err)
+		assert.NotNil(t, stats)
+
+		store, err := newStowRawStore(&Config{
+			Stow: &StowConfig{
+				Kind: local.Kind,
+				Config: map[string]string{
+					local.ConfigKeyPath: tmpDir,
+				},
+			},
+			InitContainer: "tmp",
+		}, testScope.NewSubScope("y"))
+
+		assert.NoError(t, err)
+		assert.NotNil(t, store)
+
+		stats, err = os.Stat(filepath.Join(tmpDir, "tmp"))
+		assert.NoError(t, err)
+		if assert.NotNil(t, stats) {
+			assert.True(t, stats.IsDir())
+		}
+	})
+
+	t.Run("missing init container", func(t *testing.T) {
+		testScope := promutils.NewTestScope()
+		tmpDir, err := ioutil.TempDir("", "stdlib_local")
+		assert.NoError(t, err)
+
+		stats, err := os.Stat(tmpDir)
+		assert.NoError(t, err)
+		assert.NotNil(t, stats)
+
+		store, err := newStowRawStore(&Config{
+			Stow: &StowConfig{
+				Kind: local.Kind,
+				Config: map[string]string{
+					local.ConfigKeyPath: tmpDir,
+				},
+			},
+		}, testScope.NewSubScope("y"))
+
+		assert.Error(t, err)
+		assert.Nil(t, store)
+	})
+
+	t.Run("multi-container enabled", func(t *testing.T) {
+		testScope := promutils.NewTestScope()
+		tmpDir, err := ioutil.TempDir("", "stdlib_local")
+		assert.NoError(t, err)
+
+		stats, err := os.Stat(tmpDir)
+		assert.NoError(t, err)
+		assert.NotNil(t, stats)
+
+		store, err := newStowRawStore(&Config{
+			Stow: &StowConfig{
+				Kind: local.Kind,
+				Config: map[string]string{
+					local.ConfigKeyPath: tmpDir,
+				},
+			},
+			InitContainer:         "tmp",
+			MultiContainerEnabled: true,
+		}, testScope.NewSubScope("y"))
+
+		assert.NoError(t, err)
+		assert.NotNil(t, store)
+
+		stats, err = os.Stat(filepath.Join(tmpDir, "tmp"))
+		assert.NoError(t, err)
+		if assert.NotNil(t, stats) {
+			assert.True(t, stats.IsDir())
+		}
+	})
+}
+
+func Test_newStowRawStore(t *testing.T) {
+	type args struct {
+		cfg          *Config
+		metricsScope promutils.Scope
+	}
+	tests := []struct {
+		name    string
+		args    args
+		wantErr bool
+	}{
+		{"fail", args{&Config{}, promutils.NewTestScope()}, true},
+		{"google", args{&Config{
+			InitContainer: "flyte",
+			Stow: &StowConfig{
+				Kind: google.Kind,
+				Config: map[string]string{
+					google.ConfigProjectId: "x",
+					google.ConfigScopes:    "y",
+				},
+			},
+		}, promutils.NewTestScope()}, true},
+		{"minio", args{&Config{
+			Type:          TypeMinio,
+			InitContainer: "some-container",
+			Connection: ConnectionConfig{
+				Endpoint: config.URL{URL: utils.MustParseURL("http://minio:9000")},
+			},
+		}, promutils.NewTestScope()}, true},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := newStowRawStore(tt.args.cfg, tt.args.metricsScope)
+			if tt.wantErr {
+				assert.Error(t, err, "newStowRawStore() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			assert.NotNil(t, got, "Expected rawstore, found nil!")
+		})
+	}
 }
